requirements.txt
tinkoff-investments
pytz
pandas
numpy
pyyaml
fastapi
uvicorn
catboost
aiohttp